requirements.txt
PySide6>=6.6.0
fpdf>=1.7.2
mysql-connector-python>=8.0.0
matplotlib>=3.8.0


